pyproject.toml
[project]
name = "ai-evals-course-202507"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.57.1",
    "braintrust>=0.2.0",
    "contextkit>=0.0.8",
    "dotenv-python>=0.0.1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "instructor>=1.9.2",
    "judgy>=0.1.0",
    "litellm>=1.74.8",
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "openai>=1.95.0",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rank-bm25>=0.2.2",
    "rich>=14.1.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "tavily-python>=0.7.10",
    "thefuzz>=0.22.1",
    "tqdm>=4.67.1",
    "typing-extensions>=4.14.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nbqa>=1.7.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "nbconvert>=7.16.6",
]

[tool.ruff]
line-length = 140


[tool.ruff.lint]
select = ["I"] # Enable import sorting

[tool.ruff.lint.isort]
# Keep local imports at the bottom (this is default behavior)
known-first-party = ["src"]
lines-between-types = 1
force-single-line = false

[tool.nbqa.config]
ruff = "pyproject.toml"

[tool.nbqa.mutate]
ruff = 1
